#70: Refactored and enabled logging to Error Handling Controller.
